fix(dataset): reject non-positive idle-gap caps instead of silently disabling
reflow_idle_gaps treated cap_ms <= 0 as an identity transform, so
--max-idle-gap-cap-seconds 0 (described as collapsing gaps) silently
replayed the untouched original schedule. Raise ValueError on
non-positive caps (the config layer enforces > 0), round sub-ms caps up
to 1 ms instead of int-truncating them to a collapse-everything 0, and
keep the loader's cap as float ms.

+++ src/aiperf/dataset/loader/baseten_trace.py
@@ -79,8 +79,10 @@ class BasetenTraceDatasetLoader(BaseTraceDatasetLoader[BasetenTrace]):
             dataset, "trace_session_sample_ratio", None
         )
         gap_cap_s = getattr(dataset, "max_idle_gap_cap_seconds", None)
+        # Keep the cap as float ms; int-truncation would turn a sub-ms cap
+        # into 0 and reflow_idle_gaps rejects non-positive caps.
         self._max_idle_gap_cap_ms = (
-            int(gap_cap_s * 1000) if gap_cap_s is not None else None
+            gap_cap_s * 1000 if gap_cap_s is not None else None
         )
         self._delay_cap = DelayCapTracker(
             cap_seconds=getattr(dataset, "inter_turn_delay_cap_seconds", None)
